requirements.txt
autocommand==2.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
idna==3.4
inflect==6.0.0
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
more-itertools==9.0.0
portend==3.1.0
pydantic==1.10.2
pyhumps==3.7.3
python-switchbot==2.2.2
pytz==2022.5
requests==2.28.1
six==1.16.0
tempora==5.0.2
typing_extensions==4.4.0
urllib3==1.26.12
zc.lockfile==2.0
